pyproject.toml
[project]
name = "auth-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "passlib[argon2]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
